# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_transport_http_billing.BillingAccountDataResponse:
    properties:
      account:
        $ref: '#/definitions/internal_transport_http_billing.BillingAccountResponse'
    type: object
...
  internal_transport_http_billing.BillingAccountResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_billing.BillingAccountDataResponse'
      errorMsg:
        type: string
    type: object
  internal_transport_http_billing.BillingConfigDataResponse:
    properties:
      config:
        $ref: '#/definitions/internal_transport_http_billing.BillingConfigResponse'
    type: object
  internal_transport_http_billing.BillingConfigRequest:
    properties:
      mode:
        enum:
        - self
        - period
        - usage
        type: string
      prepaidAmountUSD:
        minimum: 0
        type: number
    required:
    - mode
    type: object
...
  internal_transport_http_billing.BillingConfigResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_billing.BillingConfigDataResponse'
      errorMsg:
        type: string
    type: object
  internal_transport_http_billing.BillingOverviewDataResponse:
    properties:
      overview:
        $ref: '#/definitions/internal_transport_http_billing.BillingOverviewResponse'
    type: object
...
  internal_transport_http_billing.BillingOverviewResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_billing.BillingOverviewDataResponse'
      errorMsg:
        type: string
    type: object
  internal_transport_http_billing.BillingPlanDataResponse:
    properties:
      plan:
        $ref: '#/definitions/internal_transport_http_billing.BillingPlanResponse'
    type: object
...
  internal_transport_http_billing.BillingPlanResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_billing.BillingPlanDataResponse'
      errorMsg:
        type: string
    type: object
...
  internal_transport_http_billing.ErrorDoc:
    properties:
      data: {}
      details: {}
      errorCode:
        type: string
      errorMsg:
        type: string
      requestId:
        type: string
    type: object
  internal_transport_http_billing.ModelPricingDataResponse:
    properties:
      modelPricing:
        $ref: '#/definitions/internal_transport_http_billing.ModelPricingResponse'
    type: object
  internal_transport_http_billing.ModelPricingListResponseDoc:
    properties:
      data:
        properties:
          results:
            items:
              $ref: '#/definitions/internal_transport_http_billing.ModelPricingResponse'
            type: array
          total:
            type: integer
        type: object
      errorMsg:
        type: string
    type: object
  internal_transport_http_billing.ModelPricingResponse:
    properties:
      cacheReadNanousdPerMTokens:
        type: integer
      cacheReadUSDPerMTokens:
        type: number
      cacheWriteNanousdPerMTokens:
        type: integer
      cacheWriteUSDPerMTokens:
        type: number
      callNanousdPerCall:
        type: integer
      callUSDPerCall:
        type: number
      createdAt:
        type: string
      currency:
        type: string
      durationNanousdPerSecond:
        type: integer
      durationUSDPerSecond:
        type: number
      id:
        type: integer
      inputNanousdPerMTokens:
        type: integer
      inputUSDPerMTokens:
        type: number
      isFree:
        type: boolean
      modelIcon:
        type: string
      modelVendor:
        type: string
      outputNanousdPerMTokens:
        type: integer
      outputUSDPerMTokens:
        type: number
      platformModelName:
        type: string
      pricingMode:
        type: string
      tieredPricingJSON:
        type: string
      updatedAt:
        type: string
    type: object
...
  internal_transport_http_billing.PlanListResponseDoc:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_transport_http_billing.BillingPlanResponse'
        type: array
      errorMsg:
        type: string
    type: object
  internal_transport_http_billing.SubscribeRequest:
    properties:
      cycles:
        maximum: 120
        minimum: 1
        type: integer
      priceID:
        minimum: 1
        type: integer
    required:
    - priceID
    type: object
  internal_transport_http_billing.SubscribeResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_billing.SubscriptionDataResponse'
      errorMsg:
        type: string
    type: object
  internal_transport_http_billing.SubscriptionDataResponse:
    properties:
      subscription:
        $ref: '#/definitions/internal_transport_http_billing.SubscriptionResponse'
    type: object
...
  internal_transport_http_billing.UpdateBillingAccountBalanceRequest:
    properties:
      balanceUSD:
        minimum: 0
        type: number
      description:
        maxLength: 255
        type: string
    type: object
  internal_transport_http_billing.UpdateBillingPlanRequest:
    properties:
      amountUSD:
        minimum: 0
        type: number
      billingInterval:
        enum:
        - month
        - year
        - lifetime
        type: string
      currency:
        maxLength: 16
        type: string
      description:
        maxLength: 255
        type: string
      discountPercent:
        maximum: 100
        minimum: 0
        type: integer
      name:
        maxLength: 64
        minLength: 1
        type: string
      periodCreditUSD:
        minimum: 0
        type: number
    required:
    - billingInterval
    - name
    type: object
  internal_transport_http_billing.UpsertModelPricingRequest:
    properties:
      cacheReadUSDPerMTokens:
        minimum: 0
        type: number
      cacheWriteUSDPerMTokens:
        minimum: 0
        type: number
      callUSDPerCall:
        minimum: 0
        type: number
      currency:
        maxLength: 16
        type: string
      durationUSDPerSecond:
        minimum: 0
        type: number
      inputUSDPerMTokens:
        minimum: 0
        type: number
      isFree:
        type: boolean
      outputUSDPerMTokens:
        minimum: 0
        type: number
      platformModelName:
        maxLength: 128
        type: string
      pricingMode:
        enum:
        - token
        - call
        - duration
        - tiered
        type: string
      tieredPricingJSON:
        maxLength: 20000
        type: string
    required:
    - platformModelName
    type: object
  internal_transport_http_billing.UsageDailyListResponseDoc:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_transport_http_billing.UsageDailyResponse'
        type: array
      errorMsg:
        type: string
    type: object
...
  internal_transport_http_billing.UsageLedgerListResponseDoc:
    properties:
      data:
        properties:
          results:
            items:
              $ref: '#/definitions/internal_transport_http_billing.UsageLedgerResponse'
            type: array
          total:
            type: integer
        type: object
      errorMsg:
        type: string
    type: object
...
  internal_transport_http_billing.UsageMonthlyListResponseDoc:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_transport_http_billing.UsageMonthlyResponse'
        type: array
      errorMsg:
        type: string
    type: object
...
info:
  contact: {}
  description: DOUB Chat 后端 API 文档
  title: DOUB Chat API
  version: 0.1.0
paths:
...
  /admin/billing/accounts/{user_id}/balance:
    patch:
      consumes:
      - application/json
      description: 设置指定用户的按量计费余额，金额单位为美元
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: 余额
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_billing.UpdateBillingAccountBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.BillingAccountResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员设置用户按量余额
      tags:
      - admin-billing
  /admin/billing/config:
    get:
      consumes:
      - application/json
      description: 查询当前全局计费模式
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.BillingConfigResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员查询计费配置
      tags:
      - admin-billing
    patch:
      consumes:
      - application/json
      description: 更新当前全局计费模式
      parameters:
      - description: 计费配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_billing.BillingConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.BillingConfigResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员更新计费配置
      tags:
      - admin-billing
  /admin/billing/model-prices:
    get:
      consumes:
      - application/json
      description: 按平台模型名查询模型按量计费配置
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ModelPricingListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员查询模型按量单价
      tags:
      - admin-billing
    put:
      consumes:
      - application/json
      description: 按平台模型名创建或更新模型按量计费配置，金额单位为美元
      parameters:
      - description: 模型单价
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_billing.UpsertModelPricingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ModelPricingDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员保存模型按量单价
      tags:
      - admin-billing
  /admin/billing/plans/{id}:
    patch:
      consumes:
      - application/json
      description: 更新周期套餐基础配置与默认价格
      parameters:
      - description: 套餐ID
        in: path
        name: id
        required: true
        type: integer
      - description: 套餐配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_billing.UpdateBillingPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.BillingPlanResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员更新周期套餐
      tags:
      - admin-billing
...
  /billing/account:
    get:
      consumes:
      - application/json
      description: 查询当前用户按量余额
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.BillingAccountResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 获取按量计费账户
      tags:
      - billing
  /billing/overview:
    get:
      consumes:
      - application/json
      description: 查询当前计费方式、周期额度或按量余额
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.BillingOverviewResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 获取当前用户计费概览
      tags:
      - billing
...
  /billing/plans:
    get:
      consumes:
      - application/json
      description: 查询所有启用的订阅套餐及价格
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.PlanListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 获取订阅套餐
      tags:
      - billing
  /billing/subscriptions:
    post:
      consumes:
      - application/json
      description: 为当前用户创建或替换订阅
      parameters:
      - description: 订阅参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_billing.SubscribeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.SubscribeResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 创建订阅
      tags:
      - billing
  /billing/usage:
    get:
      consumes:
      - application/json
      description: 查询当前用户的每日用量与费用
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索模型
        in: query
        name: query
        type: string
      - description: 状态筛选：free/billable
        in: query
        name: status
        type: string
      - description: 排序：newest/oldest/tokens_desc/cost_desc/latency_desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.UsageLedgerListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 查询用量账单
      tags:
      - billing
  /billing/usage/daily:
    get:
      consumes:
      - application/json
      description: 查询当前用户按日期聚合的用量与费用
      parameters:
      - description: 天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.UsageDailyListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 查询每日用量
      tags:
      - billing
  /billing/usage/monthly:
    get:
      consumes:
      - application/json
      description: 查询当前用户按月份聚合的用量与费用
      parameters:
      - description: 月份数量，默认近 12 个月
        in: query
        name: months
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_billing.UsageMonthlyListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_billing.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 查询月度用量
      tags:
      - billing
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"